refactor(codex): split handle_orchestration
(cherry picked from commit 83b07ba1397a06381f2bb62c3cafc8b78380e2fb)
